+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/comment/adapter/CommentAdapter.kt
@@ -1,16 +1,15 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.adapter
+package com.digginroom.digginroom.feature.room.comment.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import com.digginroom.digginroom.feature.room.customview.roominfo.comment.dialog.listener.ShowCommentMenuListener
 import com.digginroom.digginroom.model.CommentModel
 
-class CommentAdapter(private val menuClickListener: ShowCommentMenuListener) :
+class CommentAdapter(private val menuClickListener: (CommentModel) -> Unit) :
     ListAdapter<CommentModel, CommentViewHolder>(CommentDiffUtilCallback()) {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CommentViewHolder {
         return CommentViewHolder.of(parent) {
-            menuClickListener.show(getItem(it))
+            menuClickListener(getItem(it))
         }
     }
