+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/comment/adapter/CommentDiffUtilCallback.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.adapter
+package com.digginroom.digginroom.feature.room.comment.adapter
 
 import androidx.recyclerview.widget.DiffUtil
 import com.digginroom.digginroom.model.CommentModel
