+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/comment/adapter/CommentViewHolder.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.adapter
+package com.digginroom.digginroom.feature.room.comment.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
