+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/comment/uistate/CommentUiState.kt
@@ -1,0 +1,7 @@
+package com.digginroom.digginroom.feature.room.comment.uistate
+
+import com.digginroom.digginroom.feature.room.comment.uistate.state.CommentState
+
+data class CommentUiState(
+    val state: CommentState
+)
