--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfo/comment/dialog/CommentUiState.kt
@@ -1,7 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.dialog
-
-import com.digginroom.digginroom.model.CommentModel
-
-data class CommentUiState(
-    val comments: List<CommentModel>
-)
