--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfo/comment/dialog/listener/CommentMenuEvent.kt
@@ -1,6 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.dialog.listener
-
-interface CommentMenuEvent {
-    fun update()
-    fun delete()
-}
