--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfo/comment/dialog/listener/ShowCommentMenuListener.kt
@@ -1,7 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfo.comment.dialog.listener
-
-import com.digginroom.digginroom.model.CommentModel
-
-fun interface ShowCommentMenuListener {
-    fun show(comment: CommentModel)
-}
