+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/RoomPlayer.kt
@@ -1,6 +1,6 @@
 package com.digginroom.digginroom.feature.room.customview.roomplayer
 
-import com.digginroom.digginroom.feature.room.customview.roominfo.RoomInfoUiState
+import com.digginroom.digginroom.feature.room.roominfo.RoomInfoUiState
 import com.dygames.roompager.Adapter
 
 interface RoomPlayer : Adapter.ViewHolder {
